requirements.txt
# Bili-Cortex 项目依赖

# 视频下载和音频提取
yt-dlp>=2024.1.0

# 语音识别和转录
faster-whisper>=1.0.0
torch>=2.0.0

# 数据处理
numpy>=1.24.0

# 配置管理
pyyaml>=6.0

# 测试依赖
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0